Refactored, tests, and README updated (#46)
